Refactor analyzer code
