+++ tests/analyzer.test.ts
@@ -27,24 +27,24 @@ export const tests = [
   //   hmsCount: 0,
   //   googlePermissionsCount: 0,
   // },
-  // {
-  //   apk: {
-  //     packageName: 'com.megaache.trackingsdks',
-  //     filePath: sampleApk,
-  //     uploadDate: 'mar 01, 2023',
-  //   },
-  //   gmsCount: 0,
-  //   hmsCount: 3,
-  // },
   {
     apk: {
-      packageName: 'com.aswat.carrefouruae',
-      filePath: path.join(__dirname, 'samples', 'com.aswat.carrefouruae.apk'),
-      uploadDate: 'April 01, 2023',
+      packageName: 'com.megaache.trackingsdks',
+      filePath: sampleApk,
+      uploadDate: 'mar 01, 2023',
     },
-    gmsCount: 6,
-    hmsCount: 0,
+    gmsCount: 0,
+    hmsCount: 3,
   },
+  // {
+  //   apk: {
+  //     packageName: 'com.aswat.carrefouruae',
+  //     filePath: path.join(__dirname, 'samples', 'com.aswat.carrefouruae.apk'),
+  //     uploadDate: 'April 01, 2023',
+  //   },
+  //   gmsCount: 6,
+  //   hmsCount: 0,
+  // },
 
   // {
   //   apk: {
